Don't set tunnel_types on VLAN network type
The current code sets tunnel types for the neutron agent
even when the network type is 'vlan'. This results in the
agent not starting because 'vlan' is not a valid tunnel type.
This patch changes it so tunnel_types is not populated when
vlans are used.

This bug just now surfaced because of commit
8feaf6c9516094df58df84479d73779e87a79264 requiring Q_ML2_TENANT_NETWORK_TYPE
to be set to avoid being set to the 'local' type.

Conflicts:
	lib/neutron_plugins/ml2

Change-Id: If2bc57b36dad2bfb34df573581acce176604812e
(cherry picked from commit e3cfbf1580d4e0b77191a2f4bcb61bde9f8d9815)